Remove unnecessary interface
ISSUE: LIS-53
