+++ src/BusinessLogic/Domain/PromotionalWidgets/Services/WidgetSettingsService.php
@@ -8,6 +8,7 @@ use SeQura\Core\BusinessLogic\Domain\CountryConfiguration\Services\CountryConfig
 use SeQura\Core\BusinessLogic\Domain\PaymentMethod\Exceptions\PaymentMethodNotFoundException;
 use SeQura\Core\BusinessLogic\Domain\PaymentMethod\Models\SeQuraPaymentMethod;
 use SeQura\Core\BusinessLogic\Domain\PaymentMethod\Services\PaymentMethodsService;
+use SeQura\Core\BusinessLogic\Domain\PromotionalWidgets\MiniWidgetMessagesProvider;
 use SeQura\Core\BusinessLogic\Domain\PromotionalWidgets\Models\ValidateAssetsKeyRequest;
 use SeQura\Core\BusinessLogic\Domain\PromotionalWidgets\Models\Widget;
 use SeQura\Core\BusinessLogic\Domain\PromotionalWidgets\Models\WidgetInitializer;
@@ -15,7 +16,6 @@ use SeQura\Core\BusinessLogic\Domain\PromotionalWidgets\Models\WidgetSettings;
 use SeQura\Core\BusinessLogic\Domain\PromotionalWidgets\ProxyContracts\WidgetsProxyInterface;
 use SeQura\Core\BusinessLogic\Domain\PromotionalWidgets\RepositoryContracts\WidgetSettingsRepositoryInterface;
 use SeQura\Core\BusinessLogic\Domain\PromotionalWidgets\WidgetConfiguratorContracts\WidgetConfiguratorInterface;
-use SeQura\Core\BusinessLogic\Domain\PromotionalWidgets\WidgetConfiguratorContracts\MiniWidgetMessagesProviderInterface;
 use SeQura\Core\Infrastructure\Http\Exceptions\HttpRequestException;
 
 /**
@@ -53,10 +53,6 @@ class WidgetSettingsService
      * @var WidgetConfiguratorInterface
      */
     protected $widgetConfigurator;
-    /**
-     * @var MiniWidgetMessagesProviderInterface
-     */
-    protected $miniWidgetMessagesProvider;
 
     /**
      * @param WidgetSettingsRepositoryInterface $widgetSettingsRepository
@@ -65,7 +61,6 @@ class WidgetSettingsService
      * @param ConnectionService $connectionService
      * @param WidgetsProxyInterface $widgetsProxy
      * @param WidgetConfiguratorInterface $widgetConfigurator
-     * @param MiniWidgetMessagesProviderInterface $miniWidgetMessagesProvider
      */
     public function __construct(
         WidgetSettingsRepositoryInterface $widgetSettingsRepository,
@@ -73,8 +68,7 @@ class WidgetSettingsService
         CountryConfigurationService $countryConfigService,
         ConnectionService $connectionService,
         WidgetsProxyInterface $widgetsProxy,
-        WidgetConfiguratorInterface $widgetConfigurator,
-        MiniWidgetMessagesProviderInterface $miniWidgetMessagesProvider
+        WidgetConfiguratorInterface $widgetConfigurator
     ) {
         $this->widgetSettingsRepository = $widgetSettingsRepository;
         $this->paymentMethodsService = $paymentMethodsService;
@@ -82,7 +76,6 @@ class WidgetSettingsService
         $this->connectionService = $connectionService;
         $this->widgetsProxy = $widgetsProxy;
         $this->widgetConfigurator = $widgetConfigurator;
-        $this->miniWidgetMessagesProvider = $miniWidgetMessagesProvider;
     }
 
     /**
@@ -270,8 +263,8 @@ class WidgetSettingsService
             $filteredMethod->getMaxAmount() ?? 0,
             '',
             '',
-            $this->miniWidgetMessagesProvider->getMessage() ?? '',
-            $this->miniWidgetMessagesProvider->getBelowLimitMessage() ?? ''
+            $this->getMiniWidgetMessage() ?? '',
+            $this->getMiniWidgetBelowLimitMessage() ?? ''
         );
     }
 
@@ -473,5 +466,37 @@ class WidgetSettingsService
         }
 
         return $widgetSupportedProducts;
+    }
+
+    /**
+     * Returns mini widget message according to current country
+     *
+     * @return string|null
+     */
+    public function getMiniWidgetMessage(): ?string
+    {
+        return MiniWidgetMessagesProvider::MINI_WIDGET_MESSAGE[$this->getCountryCode()] ?? null;
+    }
+
+    /**
+     * Returns mini widget below limit message according to current country
+     *
+     * @return string|null
+     */
+    public function getMiniWidgetBelowLimitMessage(): ?string
+    {
+        return MiniWidgetMessagesProvider::MINI_WIDGET_BELOW_LIMIT_MESSAGE[$this->getCountryCode()] ?? null;
+    }
+
+    /**
+     * Returns current country code
+     *
+     * @return string
+     */
+    protected function getCountryCode(): string
+    {
+        $locale = $this->widgetConfigurator->getLocale();
+
+        return substr($locale, strpos($locale, '_') + 1);
     }
 }
